test(detox): split interactive dialog plugin suite into four specs
Extract shared setup/selectors into interactive_dialog_test_helper and
split basic, select, text-flow, and datetime cases into separate files.
